stylistic changes, implemenatation updated branches, better web display, improved folder layout
