+++ src/App.js
@@ -1,6 +1,5 @@
-import { MovieTile } from "./MovieTile";
 import Header from "./common/Header";
-import PopularMovies from "./features/PopularMovies";
+import PopularMovies from "./features/movies/PopularMovies";
 import { Main } from "./common/Main";
 
 function App() {
