+++ src/GlobalStyle.js
@@ -13,7 +13,6 @@ export const GlobalStyle = createGlobalStyle`
         font-weight: 400;
         background: ${({ theme }) => theme.colors.whisper};
         color: ${({ theme }) => theme.colors.black};
-        display: flex;
         justify-content: center;
     }
 `;
