+++ src/index.js
@@ -1,13 +1,13 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-import { GlobalStyle } from './GlobalStyle';
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App";
+import reportWebVitals from "./reportWebVitals";
+import { GlobalStyle } from "./GlobalStyle";
 import { theme } from "./theme";
-import { ThemeProvider } from 'styled-components';
-import { Normalize } from 'styled-normalize';
+import { ThemeProvider } from "styled-components";
+import { Normalize } from "styled-normalize";
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
+const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <ThemeProvider theme={theme}>
     <React.StrictMode>
@@ -17,6 +17,5 @@ root.render(
     </React.StrictMode>
   </ThemeProvider>
 );
-
 
 reportWebVitals();
